#85 update `f` docs
