+++ builtins/core/io/f_test.go
@@ -1,0 +1,18 @@
+package io
+
+import (
+	"testing"
+
+	"github.com/lmorg/murex/test"
+)
+
+func TestLsF(t *testing.T) {
+	tests := []test.MurexTest{
+		// f
+		{
+			Block:  "f: +f",
+			Stdout: "README.md",
+		},
+	}
+	test.RunMurexTestsRx(tests, t)
+}
