+++ builtins/core/io/f_windows.go
