+++ docs/commands/f.md
@@ -2,11 +2,21 @@
 
 ## Command Reference: `f`
 
-> Lists objects (eg files) in the current working directory
+> Lists or filters file system objects (eg files)
 
 ## Description
 
-Lists objects (eg files) in the current working directory.
+`f` lists or filters lists of file system objects, such as files, directories,
+etc. `f` provides a quick way to output file system items that confirm to
+specific criteria.
+
+You define this criteria by using +flags (eg `+s` for all symlinks) and
+optionally then restricting that criteria with -flags (eg `-x` to remove all
+executable items). All flags supported as `+` are also supported as a `-`.
+
+By default `f` will return no results. You need to include +flags.
+
+Output is a JSON array as this format preserves whitespace in file names.
 
 ## Usage
 
@@ -22,29 +32,77 @@ Return only directories:
     
 Return file and directories but exclude symlinks:
 
-    f: +d +f -s
+    f: +fd -s
     
-Compare list against files (eg created by `g`) against conditions set by `f`:
+Filter out files in a list (eg created by `g`) using conditions set by `f`:
 
-    g /* -> f +f
+    g: '*.go' -> f: +f
+    
+    rx: '\.(txt|md)' -> f: +fw
 
 ## Flags
 
+* `+`
+    include files (pair this with any other flag apart from `-`)
+* `-`
+    exclude files (pair this with any other flag apart from `+`)
+* `?`
+    irregular files (nothing else is known about these files)
+* `D`
+    regular directories
+* `E`
+    other read permissions
+* `F`
+    regular files (exc symlinks, devices, sockets, named pipes, etc)
+* `Q`
+    other write permissions
+* `R`
+    user read permissions
+* `S`
+    sockets
+* `W`
+    user write permissions
+* `X`
+    user execute permissions
+* `Z`
+    other execute permissions
+* `b`
+    block devices
+* `c`
+    character devices
 * `d`
-    directories
+    all directories (inc symlinks)
+* `e`
+    group read permissions
 * `f`
-    files
+    all files (inc symlinks, devices, sockets, name pipes, etc)
+* `l`
+    symlinks
+* `p`
+    POSIX named pipes (FIFO)
+* `q`
+    group write permissions
+* `r`
+    read permissions (user, group, or other)
 * `s`
-    symlinks (automatically included with files and directories)
+    symlinks
+* `w`
+    write permissions (user, group, or other)
+* `x`
+    execute permissions (user, group, or other)
+* `z`
+    group execute permissions
 
 ## Detail
 
-By default `f` will return no results. It is then your responsibility to select
-which types of objects to be included or excluded from the results.
+`+` flags are always matched first. Then the `-` flags are used to filter out
+any matches from the `+` flags.
 
 ## See Also
 
 * [commands/`g`](../commands/g.md):
   Glob pattern matching for file system objects (eg *.txt)
+* [types/`json` ](../types/json.md):
+  JavaScript Object Notation (JSON) (primitive)
 * [commands/`rx`](../commands/rx.md):
   Regexp pattern matching for file system objects (eg '.*\.txt')
